Don't assume template sources are on the local filesystem. Instead, the StandardTemplateManager is given an instance of the new interface TemplateSource, which is responsible for determining if the source to a template is available, when it was last modified, an "external name" of the template source, and providing a Reader of the contents of that template.
We provide one implementation of TemplateSource, the
FileTemplateSource, which (surprise) looks on the local file system
under a designated source dir for templates.

NOTE: In particular, this implementation first looks for templates
with the extension '.jamon', then '.jam', and finally with no
extension.

JamonParseException can no longer assume it's
parsing a file, so it's given the "filename" in its constructor.

These changes should be externally invisible, unless there are files
ending in '.jam' or '.jamon' in the template source directories.


git-svn-id: svn://svn.code.sf.net/p/jamon/code/trunk@526 6afd1a24-8710-0410-9bc4-c69fca332c4c